Readme for memtest.txt
This readme contains everything you need to run my me test algo.  Since it’s written in Haskell, it has excellent performance when testing environments with over 1TB of RAM.  Speed of executing will be limited to the processors it’s running on.  This will work for Intel, AMD and ARM environments.